move symbolic link source ti `tests/utils`
