+++ tests/server/fixtures.py
@@ -1,1 +1,1 @@
-fixtures.py
+utils/fixtures.py
